Rename config to customization to clear up intent
it is not a configuration but a customization of the layer import
move tests into separate files instead of one big file
