--- src/config/defaults.js
@@ -1,7 +1,0 @@
-import dateTime from '../config/datetime';
-import layerName from '../config/layername';
-import geogig from '../config/geogig';
-import permisions from '../config/permissions';
-
-const configDefaults = { "layerName": layerName, "dateTime": dateTime, permissions: permisions };
-export default configDefaults;
