+++ src/customization/defaults.js
@@ -1,0 +1,7 @@
+import dateTime from '../customization/datetime';
+import layerName from '../customization/layername';
+import geogig from '../customization/geogig';
+import permissions from '../customization/permissions';
+
+const customizationDefaults = { "layerName": layerName, "dateTime": dateTime, "permissions": permissions };
+export default customizationDefaults;
